Moving Arrays to Constants File
I moved the arrays for SidebarPageData and SidbarSettingsLogoutData into the Constants.js file, then used it by importing from that file.
